+++ src/main/java/com/vwo/enums/UrlEnum.java
@@ -19,7 +19,9 @@ public enum UrlEnum {
     EVENTS("/events/t"),
     ATTRIBUTE_CHECK("/check-attribute"),
     GET_USER_DATA("/get-user-details"),
-    BATCH_EVENTS("/server-side/batch-events-v2");
+    BATCH_EVENTS("/server-side/batch-events-v2"),
+    SET_ALIAS("/user-alias/setUserAlias"),
+    GET_ALIAS("/user-alias/getAliasUserId");
 
     private final String url;