+++ src/main/java/com/vwo/packages/network_layer/models/RequestModel.java
@@ -135,6 +135,8 @@ public class RequestModel {
     }
     if (headers != null) {
       options.put("headers", headers);
+    } else {
+      headers = new HashMap<>();
     }
 
     if (method != null) {